schema: only delete old config files if the came from the controller

+++ renderer/ucentral.uc
@@ -60,8 +60,8 @@ try {
 		fs.symlink(ARGV[2], '/etc/ucentral/ucentral.active');
 
 		set_service_state(true);
-
-		fs.unlink(old_config);
+		if (split(old_config, ".")[1] == "/etc/ucentral/ucentral")
+			fs.unlink(old_config);
 	} else {
 		error = 1;
 	}
